requirements.txt
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-googledrive-storage
django-heroku
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
gunicorn
heroku
mysql-connector-python
PyMySQL
pyparsing
sqlparse
whitenoise